requirements.txt
alembic==1.4.2
aniso8601==8.0.0
attrs==20.1.0
click==7.1.2
flasgger==0.9.5
Flask==1.1.2
Flask-Admin==1.5.6
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.4
gunicorn==20.0.4
importlib-metadata==1.7.0
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
Mako==1.1.3
MarkupSafe==1.1.1
marshmallow==3.7.1
mistune==0.8.4
psycopg2==2.8.5
psycopg2-binary==2.8.5
pyrsistent==0.16.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
pytz==2020.1
PyYAML==5.3.1
six==1.15.0
SQLAlchemy==1.3.19
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.1.0
